+++ src/pygpsclient/app.py
@@ -124,22 +124,20 @@ class App(Frame):
         # user-defined serial port can be passed as environment variable
         # or command line keyword argument
         self._configfile = kwargs.pop("config", CONFIGFILE)
-        self.user_port = kwargs.pop("userport", getenv("PYGPSCLIENT_USERPORT", ""))
-        self.spartn_user_port = kwargs.pop(
+        user_port = kwargs.pop("userport", getenv("PYGPSCLIENT_USERPORT", ""))
+        spartn_user_port = kwargs.pop(
             "spartnport", getenv("PYGPSCLIENT_SPARTNPORT", "")
         )
-        self.mqapikey = kwargs.pop("mqapikey", getenv("MQAPIKEY", ""))
-        self.mqttclientid = kwargs.pop("mqttclientid", getenv("MQTTCLIENTID", ""))
-        self.mqttclientregion = kwargs.pop(
+        mqapikey = kwargs.pop("mqapikey", getenv("MQAPIKEY", ""))
+        mqttclientid = kwargs.pop("mqttclientid", getenv("MQTTCLIENTID", ""))
+        mqttclientregion = kwargs.pop(
             "mqttclientregion", getenv("MQTTCLIENTREGION", "eu")
         )
-        self.mqttclientmode = int(
+        mqttclientmode = int(
             kwargs.pop("mqttclientmode", getenv("MQTTCLIENTMODE", "0"))
         )
-        self.ntripcaster_user = kwargs.pop(
-            "ntripuser", getenv("PYGPSCLIENT_USER", "anon")
-        )
-        self.ntripcaster_password = kwargs.pop(
+        ntripcaster_user = kwargs.pop("ntripuser", getenv("PYGPSCLIENT_USER", "anon"))
+        ntripcaster_password = kwargs.pop(
             "ntrippassword", getenv("PYGPSCLIENT_USER", "password")
         )
 
@@ -168,7 +166,6 @@ class App(Frame):
         self.rtcm_handler = RTCM3Handler(self)
         self.ntrip_handler = GNSSNTRIPClient(self, verbosity=0)
         self.spartn_handler = GNSSMQTTClient(self, verbosity=0)
-        self.saved_config = {}
         self._conn_status = DISCONNECTED
         self._rtk_conn_status = DISCONNECTED
         self._socket_thread = None
@@ -179,22 +176,30 @@ class App(Frame):
         # Load configuration from file if it exists
         self._colortags = []
         self._ubxpresets = []
+        self.saved_config = {}
         (_, config, configerr) = self.file_handler.load_config(self._configfile)
         if configerr == "":  # load succeeded
             self.saved_config = config
             # update configs needed to instantiate widgets and protocol handlers
             self.update_widgets()
-        else:  # load failed - invalid json or attribute types
-            self.saved_config = {
-                "userport_s": self.user_port,
-                "spartnport_s": self.spartn_user_port,
-                "mqapikey_s": self.mqapikey,
-                "mqttclientid_s": self.mqttclientid,
-                "mqttclientregion_s": self.mqttclientregion,
-                "mqttclientmode_n": self.mqttclientmode,
-                "ntripcasteruser_s": self.ntripcaster_user,
-                "ntripcasterpassword_s": self.ntripcaster_password,
-            }
+
+        # temporarily override saved config with any command line arguments / env variables
+        if user_port != "":
+            self.saved_config["userport_s"] = user_port
+        if spartn_user_port != "":
+            self.saved_config["spartnport_s"] = spartn_user_port
+        if mqapikey != "":
+            self.saved_config["mqapikey_s"] = mqapikey
+        if mqttclientid != "":
+            self.saved_config["mqttclientid_s"] = mqttclientid
+        if mqttclientregion != "eu":
+            self.saved_config["mqttclientregion_s"] = mqttclientregion
+        if mqttclientmode != 0:
+            self.saved_config["mqttclientmode_n"] = mqttclientmode
+        if ntripcaster_user != "anon":
+            self.saved_config["ntripcasteruser_s"] = ntripcaster_user
+        if ntripcaster_password != "password":
+            self.saved_config["ntripcasterpassword_s"] = ntripcaster_password
 
         # update NTRIP and SPARTN client handlers with initial config
         self.update_NTRIP_handler()
@@ -539,15 +544,9 @@ class App(Frame):
             spartnsettings["port"] = self.saved_config.get(
                 "mqttclientport_n", SPARTN_OUTPORT
             )
-            spartnsettings["clientid"] = self.saved_config.get(
-                "mqttclientid_s", self.mqttclientid
-            )
-            spartnsettings["region"] = self.saved_config.get(
-                "mgttclientregion_s", self.mqttclientregion
-            )
-            spartnsettings["mode"] = self.saved_config.get(
-                "mgttclientmode_n", self.mqttclientmode
-            )
+            spartnsettings["clientid"] = self.saved_config.get("mqttclientid_s", "")
+            spartnsettings["region"] = self.saved_config.get("mgttclientregion_s", "eu")
+            spartnsettings["mode"] = self.saved_config.get("mgttclientmode_n", 0)
             spartnsettings["topic_ip"] = self.saved_config.get("mgttclienttopicip_b", 1)
             spartnsettings["topic_mga"] = self.saved_config.get(
                 "mgttclienttopicmga_b", 1
@@ -557,11 +556,11 @@ class App(Frame):
             )
             spartnsettings["tlscrt"] = self.saved_config.get(
                 "mgttclienttlscrt_s",
-                path.join(Path.home(), f"device-{self.mqttclientid}-pp-cert.crt"),
+                path.join(Path.home(), f"device-mqttclientid-pp-cert.crt"),
             )
             spartnsettings["tlskey"] = self.saved_config.get(
                 "mgttclienttlskey_s",
-                path.join(Path.home(), f"device-{self.mqttclientid}-pp-key.pem"),
+                path.join(Path.home(), f"device-mqttclientid-pp-key.pem"),
             )
             spartnsettings["output"] = self.spartn_inqueue
             self.spartn_handler.settings = spartnsettings